requirements.txt
streamlit
google-generativeai
PIL
python-dotenv

